wangmingyang/requirements.txt
click==6.7
diff-match-patch==20121119
Django==1.9.8
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-reversion==2.0.13
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
httplib2==0.9.2
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.2
Pillow==5.1.0
PyMySQL==0.8.0
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.7
tablib==0.12.1
tornado==4.3
unicodecsv==0.14.1
Werkzeug==0.14.1
WTForms==2.1
xlrd==1.1.0
xlwt==1.3.0
